requirements.txt
asgiref==3.7.2
Django==4.2.5
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
djangorestframework